client-v2/requirements.txt
kurigram
tgcrypto
motor
dnspython
shortzy
humanize
pytz
python-dotenv
